Prettified code

+++ CPU/inc/bcommand.h
@@ -5,7 +5,7 @@
 #pragma pack(push, 1)
 struct BinCommand {
 	uint8_t type;
-	int32_t arg1;
+	uint8_t arg1;
 	int32_t arg2;
 };
 #pragma pack(pop)
